refactor: reorganize avatar types and update imports to improve structure
