+++ apps/api/src/user/user.service.ts
@@ -5,7 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { User } from '@repo/db/entities/user';
 import { Repository } from 'typeorm';
 import { PasswordUtils } from '../common/utils/password.utils';
-import { AvatarHairStyle, AvatarMood } from '@repo/db/types/avatar';
+import { AvatarHairStyle, AvatarMood } from '@repo/db/types/user/avatar';
 
 /**
  * Service responsible for managing user data in the application.
