+++ packages/db/src/types/user/avatar.ts
